requirements.txt
requests>=2.25.0
python-dotenv>=0.15.0
google-cloud-bigquery>=2.30.0